// src/api/services/dcje/actsDCJE/actsDCJE.interface.ts
export interface GetProceduralRecordDCJEDataByActIdResponse {
    status: string;
    message: string;
    data: ProvisoryPutProceduralRecordMulti;
}
// ...
export interface ProvisoryPutProceduralRecordMulti {
    id: number;
    idProcesso: number;
    idAto: number;
    idRazaoPedido: number;
    txOrgao: string;
    txVara: string;
    txNumeroMandadoSeguranca: string;
    txAutor: string;
    txReu: string;
    nuAutores: number;
    vaTotal: number;
    dtAtualizacaoValor: string;
    idProcurador: number;
    dtPrazoProcurador: string;
    dtPrazoDCJE: string;
    txFaseProcessual: string;
    dtAjuizamento: string;
    dtCitacao: string;
    dtTransitoJulgado: string;
    txMatricula: string;
    dtAposentadoria: string;
    nuHonorariosPercentual: number;
    vaHonorariosFixos: number;
    txBaseIncidencia: string;
    dtFixacao: string;
    lsFichasProcessuaisParametrosCalculo: LsFichasProcessuaisParametrosCalculo[];
    txOrientacaoCalculo: string;
    txObservacoesGerais: string;
    idDistribuicao: number;
    dtCadastro: string;
    hrCadastro: string;
    idUsuarioCadastro: number;
    lsArquivos?: any;
    txNumeroFormatado: string;
    txTipoProcesso: string;
    idResposta: number;
    txDataResposta: string;
    vaDivergencia: number;
    txContador: string;
    isEncerrado: boolean;
    isDevolvido: boolean;
    idDevolucao: number;
}

export interface LsFichasProcessuaisParametrosCalculo {
    id: number;
    idFichaProcessual: number;
    txTipo: string;
    txIndice: string;
    txObservacao: string;
    dtInicio: string;
    dtFim: string;
}

export interface ProvisoryPutProceduralRecord {
    id?: number;
    idProcesso?: number;
    idAto?: number;
    idRazaoPedido?: number;
    txOrgao?: string;
    txVara?: string;
    txNumeroMandadoSeguranca?: string;
    txAutor?: string;
    txReu?: string;
    nuAutores?: number;
    vaTotal?: number;
    dtAtualizacaoValor?: string;
    idProcurador?: string;
    dtPrazoProcurador?: string;
    dtPrazoDCJE?: string;
    txFaseProcessual?: string;
    dtAjuizamento?: string;
    dtCitacao?: string;
    dtTransitoJulgado?: string;
    txMatricula?: string;
    dtAposentadoria?: string;
    nuHonorariosPercentual?: number;
    vaHonorariosFixos?: number;
    txBaseIncidencia?: string;
    dtFixacao?: string;
    txIndiceCorrecao?: string;
    txIndiceJuros?: string;
    txTermoJurosMora?: string;
    txObservacaoCorrecao?: string;
    txObservacaoJurosMora?: string;
    txOrientacaoCalculo?: string;
    txObservacoesGerais?: string;
    idDistribuicao?: number;
    dtCadastro?: string;
    hrCadastro?: string;
    idUsuarioCadastro: string;
    lsArquivos?: any;
    txNumeroFormatado?: string;
    txTipoProcesso?: string;
    idResposta?: number;
    txDataResposta?: string;
    vaDivergencia?: number;
    txContador?: string;
    isEncerrado?: boolean;
    isDevolvido?: boolean;
    idDevolucao?: number;
}